+++ crates/rune/src/tests/float.rs
@@ -11,7 +11,7 @@ fn test_float_fns() {
 
     let n: f64 = rune! {
         pub fn main() {
-            std::float::min(1.0, 2.0)
+            std::f64::min(1.0, 2.0)
         }
     };
     assert_eq!(n, 1.0);
@@ -25,7 +25,7 @@ fn test_float_fns() {
 
     let n: f64 = rune! {
         pub fn main() {
-            std::float::max(1.0, 2.0)
+            std::f64::max(1.0, 2.0)
         }
     };
     assert_eq!(n, 2.0);
@@ -39,7 +39,7 @@ fn test_float_fns() {
 
     let n: f64 = rune! {
         pub fn main() {
-            std::float::abs(-10.0)
+            std::f64::abs(-10.0)
         }
     };
     assert_eq!(n, 10.0);
@@ -53,7 +53,7 @@ fn test_float_fns() {
 
     let n: f64 = rune! {
         pub fn main() {
-            std::float::powf(12.0, 3.0)
+            std::f64::powf(12.0, 3.0)
         }
     };
     assert_eq!(n, 1728.0);
