+++ crates/rune/src/tests/generics.rs
@@ -95,7 +95,7 @@ fn test_sort() {
     let values: Vec<i64> = rune! {
         pub fn main() {
             let vec = [4, 3, 2, 1];
-            vec.sort::<int>();
+            vec.sort::<i64>();
             vec
         }
     };
@@ -104,7 +104,7 @@ fn test_sort() {
     let values: Vec<i64> = rune! {
         pub fn main() {
             let vec = [4, 3, 2, 1];
-            Vec::sort::<int>(vec);
+            Vec::sort::<i64>(vec);
             vec
         }
     };
@@ -113,7 +113,7 @@ fn test_sort() {
     let values: Vec<i64> = rune! {
         pub fn main() {
             let vec = [4, 3, 2, 1];
-            let f = Vec::sort::<int>;
+            let f = Vec::sort::<i64>;
             f(vec);
             vec
         }
